Making ajax requests
I made modals in store and update books and authors to function with
ajax requests

+++ app/Http/Controllers/AutorContoller.php
@@ -102,7 +102,11 @@ class AutorContoller extends Controller
      */
     public function destroy(Autor $autor)
     {
+      $auth = Autor::find($autor->id);
+      if($auth->book()->detach()){
         $autor->delete();
         return redirect()->route('autor.index');
+      }
+      return redirect()->route('autor.index');
     }
 }
